+++ qa-test-app/src/main/java/com/mapbox/navigation/qa_test_app/view/RestStopActivity.kt
@@ -1,0 +1,238 @@
+package com.mapbox.navigation.qa_test_app.view
+
+import android.annotation.SuppressLint
+import android.location.Location
+import android.os.Bundle
+import android.util.Log
+import android.view.View
+import androidx.appcompat.app.AppCompatActivity
+import com.mapbox.api.directions.v5.models.DirectionsRoute
+import com.mapbox.geojson.Point
+import com.mapbox.maps.CameraOptions
+import com.mapbox.maps.EdgeInsets
+import com.mapbox.maps.plugin.animation.CameraAnimationsPlugin
+import com.mapbox.maps.plugin.animation.MapAnimationOptions
+import com.mapbox.maps.plugin.animation.camera
+import com.mapbox.maps.plugin.locationcomponent.location
+import com.mapbox.navigation.base.options.NavigationOptions
+import com.mapbox.navigation.core.MapboxNavigation
+import com.mapbox.navigation.core.MapboxNavigationProvider
+import com.mapbox.navigation.core.replay.MapboxReplayer
+import com.mapbox.navigation.core.replay.ReplayLocationEngine
+import com.mapbox.navigation.core.replay.route.ReplayProgressObserver
+import com.mapbox.navigation.core.replay.route.ReplayRouteMapper
+import com.mapbox.navigation.core.trip.roadobject.getAllRestStops
+import com.mapbox.navigation.core.trip.roadobject.getFirstUpcomingRestStop
+import com.mapbox.navigation.core.trip.session.LocationMatcherResult
+import com.mapbox.navigation.core.trip.session.LocationObserver
+import com.mapbox.navigation.core.trip.session.RouteProgressObserver
+import com.mapbox.navigation.qa_test_app.R
+import com.mapbox.navigation.qa_test_app.databinding.LayoutActivityRestStopBinding
+import com.mapbox.navigation.qa_test_app.utils.Utils
+import com.mapbox.navigation.ui.maps.NavigationStyles
+import com.mapbox.navigation.ui.maps.location.NavigationLocationProvider
+import com.mapbox.navigation.ui.maps.route.RouteLayerConstants
+import com.mapbox.navigation.ui.maps.route.arrow.api.MapboxRouteArrowApi
+import com.mapbox.navigation.ui.maps.route.arrow.api.MapboxRouteArrowView
+import com.mapbox.navigation.ui.maps.route.arrow.model.RouteArrowOptions
+import com.mapbox.navigation.ui.maps.route.line.api.MapboxRouteLineApi
+import com.mapbox.navigation.ui.maps.route.line.api.MapboxRouteLineView
+import com.mapbox.navigation.ui.maps.route.line.model.MapboxRouteLineOptions
+import com.mapbox.navigation.ui.maps.route.line.model.RouteLine
+import com.mapbox.navigation.ui.maps.route.line.model.RouteLineResources
+
+class RestStopActivity : AppCompatActivity() {
+
+    private companion object {
+        private const val TAG = "RestStopActivity"
+    }
+
+    private val replayRouteMapper = ReplayRouteMapper()
+    private val mapboxReplayer = MapboxReplayer()
+    private val navigationLocationProvider = NavigationLocationProvider()
+
+    private val binding: LayoutActivityRestStopBinding by lazy {
+        LayoutActivityRestStopBinding.inflate(layoutInflater)
+    }
+
+    private val mapboxNavigation: MapboxNavigation by lazy {
+        MapboxNavigationProvider.create(
+            NavigationOptions.Builder(this)
+                .accessToken(Utils.getMapboxSapaAccessToken(this))
+                .locationEngine(ReplayLocationEngine(mapboxReplayer))
+                .build()
+        )
+    }
+
+    private val mapCamera: CameraAnimationsPlugin by lazy {
+        binding.mapView.camera
+    }
+
+    private val options: MapboxRouteLineOptions by lazy {
+        MapboxRouteLineOptions.Builder(this)
+            .withRouteLineResources(RouteLineResources.Builder().build())
+            .withRouteLineBelowLayerId("road-label-navigation")
+            .displayRestrictedRoadSections(true)
+            .withVanishingRouteLineEnabled(true)
+            .build()
+    }
+
+    private val routeLineView by lazy {
+        MapboxRouteLineView(options)
+    }
+
+    private val routeLineApi: MapboxRouteLineApi by lazy {
+        MapboxRouteLineApi(options)
+    }
+
+    private val routeArrowApi: MapboxRouteArrowApi by lazy {
+        MapboxRouteArrowApi()
+    }
+
+    private val routeArrowOptions by lazy {
+        RouteArrowOptions.Builder(this)
+            .withAboveLayerId(RouteLayerConstants.TOP_LEVEL_ROUTE_LINE_LAYER_ID)
+            .build()
+    }
+
+    private val routeArrowView: MapboxRouteArrowView by lazy {
+        MapboxRouteArrowView(routeArrowOptions)
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(binding.root)
+        initNavigation()
+        initStyle()
+        initListeners()
+    }
+
+    override fun onStop() {
+        super.onStop()
+        mapboxNavigation.unregisterRouteProgressObserver(routeProgressObserver)
+        mapboxNavigation.unregisterRouteProgressObserver(replayProgressObserver)
+        mapboxNavigation.unregisterLocationObserver(locationObserver)
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        routeLineApi.cancel()
+        routeLineView.cancel()
+        mapboxReplayer.finish()
+        mapboxNavigation.onDestroy()
+    }
+
+    private fun initStyle() {
+        binding.mapView.getMapboxMap().loadStyleUri(
+            NavigationStyles.NAVIGATION_DAY_STYLE
+        ) { style ->
+
+            val route = getRoute()
+            routeLineApi.setRoutes(listOf(RouteLine(route, null))) {
+                routeLineView.renderRouteDrawData(style, it)
+            }
+
+            val routeOrigin = Utils.getRouteOriginPoint(route)
+            val cameraOptions = CameraOptions.Builder().center(routeOrigin).zoom(15.0).build()
+            binding.mapView.getMapboxMap().setCamera(cameraOptions)
+        }
+    }
+
+    private fun initNavigation() {
+        binding.mapView.location.apply {
+            setLocationProvider(navigationLocationProvider)
+            enabled = true
+        }
+        mapboxNavigation.setRoutes(listOf(getRoute()))
+        mapboxNavigation.registerLocationObserver(locationObserver)
+        mapboxNavigation.registerRouteProgressObserver(replayProgressObserver)
+        mapboxReplayer.pushRealLocation(this, 0.0)
+        mapboxReplayer.playbackSpeed(1.5)
+        mapboxReplayer.play()
+    }
+
+    private val locationObserver: LocationObserver = object : LocationObserver {
+        override fun onNewRawLocation(rawLocation: Location) {
+        }
+
+        override fun onNewLocationMatcherResult(locationMatcherResult: LocationMatcherResult) {
+            navigationLocationProvider.changePosition(
+                locationMatcherResult.enhancedLocation,
+                locationMatcherResult.keyPoints,
+            )
+            updateCamera(locationMatcherResult.enhancedLocation)
+        }
+    }
+
+    private fun updateCamera(location: Location) {
+        val mapAnimationOptionsBuilder = MapAnimationOptions.Builder()
+        mapAnimationOptionsBuilder.duration(1500L)
+        mapCamera.easeTo(
+            CameraOptions.Builder()
+                .center(Point.fromLngLat(location.longitude, location.latitude))
+                .bearing(location.bearing.toDouble())
+                .zoom(15.0)
+                .padding(EdgeInsets(1000.0, 0.0, 0.0, 0.0))
+                .build(),
+            mapAnimationOptionsBuilder.build()
+        )
+    }
+
+    @SuppressLint("MissingPermission")
+    private fun initListeners() {
+        binding.startNavigation.setOnClickListener {
+            mapboxNavigation.registerRouteProgressObserver(routeProgressObserver)
+            mapboxNavigation.startTripSession()
+            binding.startNavigation.visibility = View.GONE
+            startSimulation(mapboxNavigation.getRoutes()[0])
+        }
+    }
+
+    private fun startSimulation(route: DirectionsRoute) {
+        mapboxReplayer.stop()
+        mapboxReplayer.clearEvents()
+        val replayData = replayRouteMapper.mapDirectionsRouteGeometry(route)
+        mapboxReplayer.pushEvents(replayData)
+        mapboxReplayer.seekTo(replayData[0])
+        mapboxReplayer.play()
+    }
+
+    private val routeProgressObserver = RouteProgressObserver { routeProgress ->
+        routeLineApi.updateWithRouteProgress(routeProgress) { result ->
+            binding.mapView.getMapboxMap().getStyle()?.apply {
+                routeLineView.renderRouteLineUpdate(this, result)
+            }
+        }
+
+        val arrowUpdate = routeArrowApi.addUpcomingManeuverArrow(routeProgress)
+        binding.mapView.getMapboxMap().getStyle()?.apply {
+            routeArrowView.renderManeuverUpdate(this, arrowUpdate)
+        }
+
+        val allRestStops = routeProgress.upcomingRoadObjects.getAllRestStops()
+        val firstUpcomingRestStop = routeProgress.upcomingRoadObjects.getFirstUpcomingRestStop()
+        Log.d(
+            "RestStopActivity",
+            "first upcoming restStop: " +
+                "${firstUpcomingRestStop?.restStop?.name} -- " +
+                "${firstUpcomingRestStop?.restStop?.restStopType} -- " +
+                "${firstUpcomingRestStop?.restStop?.amenities}"
+        )
+        allRestStops.forEach {
+            Log.d(
+                "RestStopActivity",
+                "all rest stops with route progress: " +
+                    "${it.restStop?.name} -- " +
+                    "${it.restStop?.restStopType} -- " +
+                    "${it.restStop?.amenities}"
+            )
+        }
+    }
+
+    private val replayProgressObserver = ReplayProgressObserver(mapboxReplayer)
+
+    private fun getRoute(): DirectionsRoute {
+        val routeAsString = Utils.readRawFileText(this, R.raw.route_with_sapa)
+        return DirectionsRoute.fromJson(routeAsString)
+    }
+}